Add TermTable::Base class
This allows us to define the :members field once and then have
subclasses tell use which headers to look for an then define a `Member`
for that table type.

I think this is a step in the right direction but the individual table
classes are now essentially just representing rows, so we can move those
out of the way as well.